2/14 VD UPDATE
Happy Valentine's Day to my computer, because that's the only thing that is getting any decent amount of attention today. Routing is in the process of being added along with cookies. Routing code now almost finished (or so it appears). Fixed a lot of things on various forms and they're almost ready to relay things server side.